backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
ollama>=0.1.7
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.1.0
alembic>=1.12.1
easyocr>=1.7.0

